Rename `operators/coord` to `operators/geometry`.
