+++ dali/operators/CMakeLists.txt
@@ -16,7 +16,7 @@ project(dali_operator CUDA CXX C)
 
 add_subdirectory(audio)
 add_subdirectory(bbox)
-add_subdirectory(coord)
+add_subdirectory(geometry)
 add_subdirectory(debug)
 add_subdirectory(decoder)
 add_subdirectory(generic)
