+++ dali/operators/geometry/coord_flip.h
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef DALI_OPERATORS_COORD_COORD_FLIP_H_
-#define DALI_OPERATORS_COORD_COORD_FLIP_H_
+#ifndef DALI_OPERATORS_GEOMETRY_COORD_FLIP_H_
+#define DALI_OPERATORS_GEOMETRY_COORD_FLIP_H_
 
 #include <string>
 #include <vector>
@@ -104,4 +104,4 @@ class CoordFlip : public Operator<Backend> {
 
 }  // namespace dali
 
-#endif  // DALI_OPERATORS_COORD_COORD_FLIP_H_
+#endif  // DALI_OPERATORS_GEOMETRY_COORD_FLIP_H_
